+++ datadog_checks_dev/datadog_checks/dev/tooling/templates/integration/check_only/{check_name}/datadog_checks/{check_name}/__about__.py
@@ -1,0 +1,2 @@
+{license_header}
+__version__ = '{starting_version}'