+++ gitlab/tests/test_unit.py
@@ -4,8 +4,7 @@
 import pytest
 from mock.mock import MagicMock
 
-from datadog_checks.base import AgentCheck, ConfigurationError
-from datadog_checks.dev.testing import requires_py2, requires_py3
+from datadog_checks.base import AgentCheck
 from datadog_checks.dev.utils import get_metadata_metrics
 from datadog_checks.gitlab.common import get_gitlab_version
 
@@ -33,7 +32,6 @@ def test_check(dd_run_check, aggregator, mock_data, gitlab_check, get_config, us
     aggregator.assert_metrics_using_metadata(get_metadata_metrics())
 
 
-@requires_py3
 def test_check_gitaly(dd_run_check, aggregator, mock_data, gitlab_check, get_config):
     from datadog_checks.gitlab.gitlab_v2 import GitlabCheckV2
 
@@ -54,14 +52,6 @@ def test_check_gitaly(dd_run_check, aggregator, mock_data, gitlab_check, get_con
         status=GitlabCheckV2.OK,
         tags=GITLAB_TAGS + CUSTOM_TAGS + ['endpoint:{}'.format(GITLAB_GITALY_PROMETHEUS_ENDPOINT)],
     )
-
-
-@requires_py2
-def test_openmetrics_with_python2(gitlab_check, get_config):
-    with pytest.raises(
-        ConfigurationError, match="This version of the integration is only available when using Python 3."
-    ):
-        gitlab_check(get_config(True))
 
 
 @pytest.mark.parametrize(
@@ -88,7 +78,6 @@ def test_get_gitlab_version_without_token():
     assert version is None
 
 
-@requires_py3
 def test_no_gitlab_url(dd_run_check, aggregator, mock_data, gitlab_check, get_config):
     config = get_config(True)
     del config['instances'][0]['gitlab_url']
@@ -97,7 +86,6 @@ def test_no_gitlab_url(dd_run_check, aggregator, mock_data, gitlab_check, get_co
     aggregator.assert_service_check('gitlab.openmetrics.health', status=AgentCheck.OK)
 
 
-@requires_py3
 def test_parse_readiness_service_checks_with_invalid_payload(
     dd_run_check, aggregator, mock_data, gitlab_check, get_config
 ):
@@ -159,7 +147,6 @@ def test_parse_readiness_service_checks_with_invalid_payload(
         ),
     ],
 )
-@requires_py3
 def test_parse_readiness_service_checks(
     dd_run_check, aggregator, mock_data, gitlab_check, get_config, service_check, expected_redis_status
 ):