+++ hyperv/tests/test_hyperv.py
@@ -4,12 +4,10 @@
 import pytest
 
 from datadog_checks.base.constants import ServiceCheck
-from datadog_checks.dev.testing import requires_py2, requires_py3
 from datadog_checks.hyperv import HypervCheck
 from datadog_checks.hyperv.metrics import DEFAULT_COUNTERS
 
 
-@requires_py3
 def test_check(aggregator, dd_default_hostname, dd_run_check):
     check = HypervCheck('hyperv', {}, [{}])
     check.hostname = dd_default_hostname
@@ -21,14 +19,6 @@ def test_check(aggregator, dd_default_hostname, dd_run_check):
     aggregator.assert_service_check(
         'hyperv.windows.perf.health', ServiceCheck.OK, count=2, tags=['server:{}'.format(dd_default_hostname)]
     )
-    _assert_metrics(aggregator)
-
-
-@requires_py2
-def test_check_legacy(aggregator, instance_refresh, dd_run_check):
-    check = HypervCheck('hyperv', {}, [instance_refresh])
-    dd_run_check(check)
-
     _assert_metrics(aggregator)