+++ postfix/tests/test_integration.py
@@ -2,7 +2,6 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import pytest
-from six import iteritems
 
 from datadog_checks.postfix import PostfixCheck
 
@@ -15,6 +14,6 @@ def test_check(aggregator):
     check = PostfixCheck('postfix', {}, [instance])
     check.check(instance)
 
-    for queue, count in iteritems(get_queue_counts()):
+    for queue, count in get_queue_counts().items():
         tags = ['instance:postfix_data', 'queue:{}'.format(queue)]
         aggregator.assert_metric('postfix.queue.size', value=count[0], tags=tags)