+++ vsphere/datadog_checks/vsphere/api.py
@@ -8,7 +8,6 @@ from typing import Any, Callable, List, TypeVar, cast  # noqa: F401
 
 from pyVim import connect
 from pyVmomi import vim, vmodl
-from six import itervalues
 
 from datadog_checks.base.log import CheckLoggingAdapter  # noqa: F401
 from datadog_checks.vsphere.config import VSphereConfig  # noqa: F401
@@ -286,7 +285,7 @@ class VSphereAPI(object):
             # at this point they are custom pyvmomi objects and the attribute keys are not resolved.
 
             attribute_keys = {x.key: x.name for x in self._fetch_all_attributes()}
-            for props in itervalues(infrastructure_data):
+            for props in infrastructure_data.values():
                 mor_attributes = []
                 if self.config.collect_property_metrics:
                     all_properties = {}